Fix Datastar interfacing with modules
The error is that $isSubmitting is not defined in the JavaScript
context. This is happening because the Datastar signals ($isSubmitting,
$loginError, etc.) are only available in the HTML context, not in the
separate JavaScript module.

The issue is that when we're in a separate JavaScript module, we don't
have direct access to the Datastar signals. We need to access them
through the Datastar API.